move diagnostics stuff to Diagnostics namespace, clean up some docs
