+++ src/Soulseek/ISoulseekClient.cs
@@ -17,8 +17,8 @@ namespace Soulseek
     using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
-    using Soulseek.Messaging.Messages;
 
     /// <summary>
     ///     A client for the Soulseek file sharing network.
