+++ src/Soulseek/Messaging/Handlers/DistributedMessageHandler.cs
@@ -14,6 +14,7 @@ namespace Soulseek.Messaging.Handlers
 {
     using System;
     using System.Threading;
+    using Soulseek.Diagnostics;
     using Soulseek.Messaging;
     using Soulseek.Messaging.Messages;
     using Soulseek.Network;
