+++ src/Soulseek/Messaging/Handlers/ServerMessageHandler.cs
@@ -15,6 +15,7 @@ namespace Soulseek.Messaging.Handlers
     using System;
     using System.Linq;
     using System.Threading;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
     using Soulseek.Messaging;
     using Soulseek.Messaging.Messages;
