+++ src/Soulseek/Network/ListenerHandler.cs
@@ -14,6 +14,7 @@ namespace Soulseek.Network
 {
     using System;
     using System.Linq;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
     using Soulseek.Messaging.Messages;
     using Soulseek.Network.Tcp;
