+++ src/Soulseek/SoulseekClientOptions.cs
@@ -17,6 +17,7 @@ namespace Soulseek
     using System.Linq;
     using System.Net;
     using System.Threading.Tasks;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
     using Soulseek.Messaging.Messages;
