+++ tests/Soulseek.Tests.Unit/Messaging/Handlers/DistributedMessageHandlerTests.cs
@@ -22,6 +22,7 @@ namespace Soulseek.Tests.Unit.Messaging.Handlers
     using System.Threading.Tasks;
     using AutoFixture.Xunit2;
     using Moq;
+    using Soulseek.Diagnostics;
     using Soulseek.Messaging;
     using Soulseek.Messaging.Handlers;
     using Soulseek.Messaging.Messages;
