+++ tests/Soulseek.Tests.Unit/Network/DistributedConnectionManagerTests.cs
@@ -23,6 +23,7 @@ namespace Soulseek.Tests.Unit.Network
     using System.Threading.Tasks;
     using AutoFixture.Xunit2;
     using Moq;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
     using Soulseek.Messaging;
     using Soulseek.Messaging.Handlers;
