+++ tests/Soulseek.Tests.Unit/Network/ListenerHandlerTests.cs
@@ -20,6 +20,7 @@ namespace Soulseek.Tests.Unit.Network
     using System.Threading.Tasks;
     using AutoFixture.Xunit2;
     using Moq;
+    using Soulseek.Diagnostics;
     using Soulseek.Messaging.Messages;
     using Soulseek.Network;
     using Soulseek.Network.Tcp;
