+++ tests/Soulseek.Tests.Unit/Network/PeerConnectionManagerTests.cs
@@ -21,6 +21,7 @@ namespace Soulseek.Tests.Unit.Network
     using System.Threading.Tasks;
     using AutoFixture.Xunit2;
     using Moq;
+    using Soulseek.Diagnostics;
     using Soulseek.Exceptions;
     using Soulseek.Messaging.Handlers;
     using Soulseek.Messaging.Messages;
